# go/go/src/ipsap/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "IPSAP Service APIs",
        "contact": {},
        "license": {},
        "version": "1.0"
    },
    "basePath": "/api/v1.0",
    "paths": {
...
        "/user/{user_seq}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "User 정보",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "User 정보",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user_seq",
                        "name": "user_seq",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {}
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "User 탈퇴(현재 로그인한 기관내에서만 탈퇴!)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "User 탈퇴 (현재 로그인한 기관내에서만 탈퇴!)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user_seq",
                        "name": "user_seq",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {}
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "나의 정보 수정",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "나의 정보 수정",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user_seq",
                        "name": "user_seq",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "param",
                        "name": "param",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.userMyInfoPatchModel"
                        }
                    }
                ],
                "responses": {
                    "200": {}
                }
            }
        },
        "/user/{user_seq}/change-password": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "User 비밀번호변경",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "User 비밀번호변경",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user_seq",
                        "name": "user_seq",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "param",
                        "name": "param",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.loginChangePwModel"
                        }
                    }
                ],
                "responses": {
                    "200": {}
                }
            }
        },
        "/user/{user_seq}/institution": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "User 탈퇴(소속된 기관 전체에서 탈퇴!)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "User 탈퇴 (소속된 기관 전체에서 탈퇴!)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user_seq",
                        "name": "user_seq",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {}
                }
            }
        },
        "/user/{user_seq}/register": {
            "patch": {
                "description": "나의 정보 등록(User가 직접입력)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "나의 정보 등록",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user_seq",
                        "name": "user_seq",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "param",
                        "name": "param",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.userRegisterModel"
                        }
                    }
                ],
                "responses": {
                    "200": {}
                }
            }
        }
    },
    "definitions": {
...
        "model.loginChangePwModel": {
            "type": "object",
            "properties": {
                "new_pw": {
                    "type": "string",
                    "example": "test"
                },
                "old_pw": {
                    "type": "string",
                    "example": "tester"
                }
            }
        },
...
        "model.userMyInfoPatchModel": {
            "type": "object",
            "properties": {
                "agree_email": {
                    "type": "integer",
                    "example": 1
                },
                "agree_pri_open": {
                    "type": "integer",
                    "example": 1
                },
                "agree_sms": {
                    "type": "integer",
                    "example": 1
                },
                "dept": {
                    "type": "string",
                    "example": "소속부서"
                },
                "edu_course_num": {
                    "type": "string",
                    "example": "0"
                },
                "edu_date": {
                    "type": "string",
                    "example": "20210101"
                },
                "edu_institution": {
                    "type": "string",
                    "example": "건국대학교"
                },
                "major_field": {
                    "type": "string",
                    "example": "수의학"
                },
                "phoneno": {
                    "type": "string",
                    "example": "021231234"
                },
                "position": {
                    "type": "string",
                    "example": "직급"
                }
            }
        },
...
        "model.userRegisterModel": {
            "type": "object",
            "properties": {
                "agree_email": {
                    "type": "integer",
                    "example": 1
                },
                "agree_pri_open": {
                    "type": "integer",
                    "example": 1
                },
                "agree_sms": {
                    "type": "integer",
                    "example": 1
                },
                "dept": {
                    "type": "string",
                    "example": "소속부서"
                },
                "edu_course_num": {
                    "type": "string",
                    "example": "0"
                },
                "edu_date": {
                    "type": "string",
                    "example": "20210101"
                },
                "edu_institution": {
                    "type": "string",
                    "example": "건국대학교"
                },
                "institution_seq": {
                    "type": "integer",
                    "example": 1
                },
                "major_field": {
                    "type": "string",
                    "example": "수의학"
                },
                "name": {
                    "type": "string",
                    "example": "홍길동"
                },
                "phoneno": {
                    "type": "string",
                    "example": "01012341234"
                },
                "position": {
                    "type": "string",
                    "example": "직급"
                },
                "pwd": {
                    "type": "string",
                    "example": "tmvmfls!00"
                },
                "tmp_key": {
                    "type": "string",
                    "example": "12345678901234561234567890123456"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "token",
            "in": "header"
        }
    }
}